Changes for the library migration in hooks
